Add the straight_line optional argument to some functions.
We really want to use navigable paths rather than straight line
distances, but straight line distances are much faster.  So we will use
straight line distances as a stopgap until we speed up the navigable
paths code enough.

We could also do two-stage processing where we do straight-line
distances first to get a lower bound on the distance so an upper bound
on the btn, and then only do the navigable paths once we know we have
an actual trade route to plot.